This commit connect from commons and not for each module anymore, this is part of a refactor, also delete Producer from core layer because is not being used anymore
