+++ management-core/src/main/java/com/ns/task/service/ProductCoreServiceImpl.java
@@ -1,9 +1,8 @@
 package com.ns.task.service;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.ns.task.config.ProducerCore;
 import com.ns.task.entities.ProductEntity;
+import com.ns.task.exceptions.ProductManagementException;
 import com.ns.task.model.ProductDTO;
 import com.ns.task.repositories.ProductRepository;
 import com.ns.task.services.ProductService;
@@ -31,8 +30,6 @@ public class ProductCoreServiceImpl implements ProductService {
     private AmqpTemplate amqpTemplate;
     private ObjectMapper objectMapper = new ObjectMapper();
 
-    @Autowired
-    ProducerCore producerCore;
 
     public ProductCoreServiceImpl() {
     }
@@ -89,14 +86,14 @@ public class ProductCoreServiceImpl implements ProductService {
             product.setMessage("Name and model of the service exists already");
             logger.debug("Getting persisted data from insert to DB {}", productEntity);
         } catch (DataIntegrityViolationException exception) {
-            //throw new ProductManagementException("Name and model of the service exists already", exception);
-            product.setComplete(false);
-            product.setMessage("Name and model of the service exists already");
-            try {
-                producerCore.produce(product);
-            } catch (JsonProcessingException e) {
-                e.printStackTrace();
-            }
+            throw new ProductManagementException("Name and model of the service exists already", exception);
+            //product.setComplete(false);
+            //product.setMessage("Name and model of the service exists already");
+            //try {
+            //producerCore.produce(product);
+            //} catch (JsonProcessingException e) {
+            //  e.printStackTrace();
+            //}
         }
         return convertToDto(productEntity);
